getDistance to always return relative distance
This will make controlling them via higher level of
abstraction easier and will make the API more consistent.

Additionally, an isAttached() method was added to the
Odometer interface so to be able to determine whether the
sensor was correctly attached, since we cannot ensure that
during compile time like in other cases.