Expect attach calls
Currently, these two expectations are there only for the coverage.
Unfortunately, due to how Arduino handles external interrupt
registrations (and particularly the lack of an std::function
implementation or the ability to attach interrupts before setup)
we need the user to call attach manually, outside the library.

Ideally, attach should not even be needed to be called
by the user as this is a sign of sequential cooupling. :/