Instantiate the Arduino runtime default argument in main header
This way we do not pollute our implementation with a seamingly
needless instatiation.
Additionally, we now define the macro to trigger the default
injection of the runtime environment on the Smartcar.h header
so to make things more transparent, as opposed to this happening
inside a cmake file for tests.