Make Smartcar library compatible with ESP32
To be able to compile the Smartcar library for ESP32
we need to provide alternative implementations for:
* analogWrite
* Servo library

To achieve this we use as submodules existing libraries
that offer the missing functionality and have adjusted
the implementation accordingly. The way it's done right
now, i.e. with ifdef statements is perhaps not the best
but it's good enough for now.