Remove hardcoded arduino runtime injection
The ArduinoRuntime was thus far injected by default in all
classes that needed it. This was achieved with a small 'hack'
which is now removed, in the name of consistency and object
orientation. This will also make it slightly simpler to move
to other platforms.

As a consequence, this is a backwards incompatible change.