Fix ESP32 power supply line to 5V
ESP32 was powered on via its VIN pin which based
on the specs of the ESP32 DOIT devkit is supposed
to handle up to 12V.
This isn't actually the case and anything above 6V
makes the regulator on the board overheat.

Note that the 5V is actually the voltage coming from
our own on-board regulator.