Support dual motors
Add support for motors connected on different pins but controlled
in the same way.
This is useful when we have two motor drivers and four
different motors where the two motors on each side of the car are
connected to the same driver but on different pins.

Fixes rosmo-robot/smartcar_shield#1